requirements.txt
streamlit==1.34.0
google-generativeai==0.8.3
python-dotenv==1.0.0